Convert login method to action
